EXLM-209: refactor  JS code per PR feedback from Ahmed. Add breakpoint to mini-toc css.
