--- util/timers.js
@@ -1,3 +1,0 @@
-'use strict';
-
-export const timers = new Map();
